Fix shader brush not initialized exception and add two improvements (#2123)
1. Add image uniform customization support from drawable resource. Also
can do image customization from bytes array but no examples and no
helper functions created. Recommended to use drawable resource to do
customization.
2. Fix shader brush not initialized exception on api level 32 where
fallback color is not set. Also updated the figma design doc to test
those cases.
3. BuildProcessor now set the shader time uniform state in the generated
code. It also adds a designFeatures to the DesignDoc annotation to
decide if the time uniform state will be set to the customization
context or not.

Fixes: #2124